Fix z17/z18 rendering bugs: multipolygon holes, line clipping, integer shifts, desktop clamping
- Render multipolygon inner rings correctly using even-odd scanline fill.
- Clip line segments independently to avoid spurious diagonals when a line
  exits and re-enters a tile; scale clip margin with line width.
- Replace signed 1 << z shifts with unsigned casts throughout engine/compiler.
- Fix desktop pan/zoom clamping to match engine viewport bounds.
- Add PBF cropper tool (compiler/tools/crop_pbf.cpp) for fast test extracts.
- Add high-zoom z18 fixture and unit tests.

+++ engine/src/mbtiles_reader.cpp
@@ -63,7 +63,7 @@ void MBTilesReader::bounds(double& w, double& s, double& e, double& n) const {
 bool MBTilesReader::read_tile(int z, int x, int slippy_y, std::vector<unsigned char>& out) {
     if (!db_) return false;
 
-    const int tms_row = (1 << z) - 1 - slippy_y;  // MBTiles stores TMS rows
+    const int tms_row = static_cast<int>((1u << z) - 1) - slippy_y;  // MBTiles stores TMS rows
     sqlite3_stmt* stmt = nullptr;
     static const char* kSql =
         "SELECT tile_data FROM tiles WHERE zoom_level=? AND tile_column=? AND tile_row=?;";
